+++ app/components/ActivityList/index.js
@@ -23,7 +23,6 @@ import { setUpdatedTime, setAppStatus, setConnection } from '../../state/app/app
 import { setActivities } from '../../state/activities/activities.actions';
 import { setScheduleUpdated } from '../../state/applets/applets.actions';
 import {
-  responseScheduleSelector,
   inProgressSelector,
 } from '../../state/responses/responses.selectors';
 
@@ -42,7 +41,6 @@ const ActivityList = ({
   scheduleUpdated,
   isConnected,
   setScheduleUpdated,
-  responseSchedule,
   inProgress,
   setActivities,
   finishedEvents,
@@ -155,7 +153,7 @@ const ActivityList = ({
         timers.current.shift()
       }
     }
-  }, [Object.keys(inProgress).length, responseSchedule, applet]);
+  }, [Object.keys(inProgress).length, applet]);
 
   useEffect(() => {
     if (appStatus) {
@@ -210,7 +208,6 @@ ActivityList.propTypes = {
   appStatus: PropTypes.bool.isRequired,
   setAppStatus: PropTypes.func.isRequired,
   setConnection: PropTypes.func.isRequired,
-  responseSchedule: PropTypes.object.isRequired,
   activityAccess: PropTypes.object.isRequired,
   appletTime: PropTypes.any.isRequired,
   inProgress: PropTypes.object.isRequired,
@@ -238,7 +235,6 @@ const mapStateToProps = (state) => {
     isConnected: connectionSelector(state),
     appletTime: state.applets.currentTime,
     activitySelectionDisabled: activitySelectionDisabledSelector(state),
-    responseSchedule: responseScheduleSelector(state),
     activityAccess: activityAccessSelector(state),
     inProgress: inProgressSelector(state),
     finishedEvents: finishedEventsSelector(state),