+++ app/components/AppletData/ItemChart.js
@@ -1,10 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { View } from 'react-native';
-import moment from 'moment';
-import i18n from 'i18next';
 
-import TokenChart from './TokenChart';
 import BaseText from '../base_text/base_text';
 
 const styles = {
@@ -53,95 +50,6 @@ class ItemChart extends React.Component {
     );
   }
 
-  renderTokenPlot() {
-    const values = {};
-    const { item, data, tokens } = this.props;
-    const { enableNegativeTokens } = item.valueConstraints;
-
-    const currentDate = new Date();
-    currentDate.setDate(currentDate.getDate() - currentDate.getDay());
-    const dayOfWeeks = i18n.t('calendar:weekdays').split('_');
-    const itemValues = item.valueConstraints.itemList.map(itemData => {
-      return {
-        name: itemData.name.en,
-        value: itemData.value,
-      }
-    });
-
-    const month = currentDate.getMonth() + 1;
-    const day = currentDate.getDate();
-    const year = currentDate.getFullYear();
-    const newDate = `${year}-${month < 10 ? '0' : ''}${month}-${day < 10 ? '0' : ''}${day}`;
-
-    if (item.additionalParams.activeCount === 0) {
-      return null;
-    }
-
-    data.forEach((val) => {
-      if (val.date >= newDate) {
-        const currentDay = dayOfWeeks[moment(val.date).day()];
-        const sum = Array.isArray(val.value)
-          ? val.value.reduce((a, b) => {
-            const bValue = !isNaN(b) ? parseInt(b) : itemValues.find(({ name }) => name === b).value;
-            if (enableNegativeTokens || bValue > 0) {
-              return a + bValue;
-            }
-            return a;
-          }, 0)
-          : val.value;
-        values[currentDay] = values[currentDay] === undefined ? sum : values[currentDay] + sum;
-      }
-    });
-
-    const dataValues = dayOfWeeks.map((dayOfWeek) => {
-      if (Object.keys(values).includes(dayOfWeek)) {
-        return {
-          name: dayOfWeek,
-          value: parseInt(values[dayOfWeek]),
-        };
-      }
-      return {
-        name: dayOfWeek,
-        value: 0,
-      };
-    });
-
-    const tokenUpdates = tokens.tokenUpdates.filter(tokenUpdate => tokenUpdate.created >= newDate).reduce((usedTokens, tokenUpdate) => {
-      const day = dayOfWeeks[moment(tokenUpdate.created).day()];
-
-      usedTokens[day] = usedTokens[day] || 0;
-      usedTokens[day] += tokenUpdate.value;
-
-      return usedTokens;
-    }, {})
-
-    const tokenHistory = {
-      tokenUpdates: dayOfWeeks.map(day => ({
-        name: day,
-        value: (tokenUpdates[day] || 0)
-      })),
-      currentBalance: Number.isInteger(tokens.cumulativeToken) ? (tokens.cumulativeToken) : 0,
-    };
-
-    const questions = item.question['en'].split("\n");
-
-    return (
-      <View style={styles.plotView}>
-        <BaseText
-          style={{
-            fontWeight: 'bold',
-            paddingBottom: 20,
-            paddingTop: 20,
-            paddingHorizontal: 20,
-          }}
-          value={questions[questions.length - 1]}
-        />
-        {/* {item.additionalParams.timelineChart} */}
-        <TokenChart item={item} data={dataValues} labels={item.additionalParams.labels} tokens={tokenHistory}/>
-      </View>
-    );
-  }
-
   // eslint-disable-next-line
   renderLinePlot() {
     const { item /* , data */ } = this.props;
@@ -171,7 +79,7 @@ class ItemChart extends React.Component {
     }
 
     const values = item.question['en'].split("\n");
-    
+
     return (
       <View style={styles.plotView}>
         <BaseText
@@ -189,12 +97,9 @@ class ItemChart extends React.Component {
   }
 
   render() {
-    const { item, type } = this.props;
+    const { item } = this.props;
     switch (item.inputType) {
       case 'radio':
-        if (type === 'TokenLogger') {
-          return this.renderTokenPlot();
-        }
         return this.renderTimelinePlot();
       case 'slider':
         return this.renderLinePlot();
@@ -208,9 +113,7 @@ class ItemChart extends React.Component {
 
 ItemChart.propTypes = {
   item: PropTypes.object.isRequired,
-  type: PropTypes.string.isRequired,
   data: PropTypes.array.isRequired,
-  tokens: PropTypes.object.isRequired,
 };
 
 export default ItemChart;