+++ app/scenes/AppletDetails/AppletDetailsComponent.js
@@ -117,9 +117,9 @@ class AppletDetailsComponent extends React.Component {
         );
       case 'about':
         return (
-          <Content>
+          <View style={{ height: '100%', flex: 1 }}>
             <AppletAbout applet={applet} />
-          </Content>
+          </View>
         );
       default:
         break;