+++ app/state/app/app.actions.js
@@ -24,6 +24,11 @@ export const setCurrentApplet = appletId => ({
   payload: appletId,
 });
 
+export const setTrailsTimerId = timerId => ({
+  type: APP_ACTIONS.SET_TRAILS_TIMER_ID,
+  payload: timerId,
+})
+
 export const setTutorialStatus = status => ({
   type: APP_ACTIONS.SET_TUTORIAL_STATUS,
   payload: status,