+++ app/state/responses/responses.actions.js
@@ -89,7 +89,12 @@ export const setCurrentScreen = (activityId, screenIndex, startTime) => ({
   },
 });
 
-export const setSchedule = schedule => ({
-  type: RESPONSES_CONSTANTS.SET_SCHEDULE,
-  payload: schedule,
+export const setLastResponseTime = responseTimes => ({
+  type: RESPONSES_CONSTANTS.SET_LAST_RESPONSE_TIME,
+  payload: responseTimes,
 });
+
+export const setCurrentBehavior = behavior => ({
+  type: RESPONSES_CONSTANTS.SET_CURRENT_BEHAVIOR,
+  payload: behavior
+})