+++ app/state/responses/responses.reducer.js
@@ -15,6 +15,7 @@ export const initialState = {
   uploadQueue: [],
   schedule: {},
   activityOpened: false,
+  currentBehavior: {},
 };
 
 const replaceAppletResponses = (state, action) => {
@@ -166,6 +167,11 @@ export default (state = initialState, action = {}) => {
           },
         },
       };
+    case RESPONSES_CONSTANTS.SET_CURRENT_BEHAVIOR:
+      return {
+        ...state,
+        currentBehavior: action.payload
+      }
     case RESPONSES_CONSTANTS.SET_RESPONSES_DOWNLOAD_PROGRESS:
       return {
         ...state,
@@ -184,10 +190,10 @@ export default (state = initialState, action = {}) => {
         ...state,
         uploadQueue: R.remove(0, 1, state.uploadQueue),
       };
-    case RESPONSES_CONSTANTS.SET_SCHEDULE:
+    case RESPONSES_CONSTANTS.SET_LAST_RESPONSE_TIME:
       return {
         ...state,
-        schedule: action.payload,
+        lastResponseTime: action.payload,
       };
     default:
       return state;