+++ app/widgets/index.js
@@ -22,4 +22,6 @@ export * from './TooltipBox';
 export * from './StackedSlider';
 export * from './AgeSelector';
 export * from './StackedRadio';
+export * from './BehaviorTracker';
+export * from './TokenSummary';
 export * from './StabilityTracker';